requirements.txt
pandas
requests
python-binance
colorama